+++ source/Energinet.Charges.Libraries/source/Energinet.DataHub.Charges.Clients.Tests/Charges/ChargesClientTests.cs
@@ -96,11 +96,11 @@ namespace Energinet.DataHub.Charges.Clients.CreateDefaultChargeLink.Tests.Charge
         [Theory]
         [InlineAutoDomainData]
         public async Task GetChargesAsync_WhenSuccess_ReturnsCharges(
-            ChargeV1Dto chargeLinkDto,
+            ChargeV1Dto chargeDto,
             Mock<IChargesClientFactory> chargesClientFactory)
         {
             // Arrange
-            var responseContent = CreateValidResponseContent(chargeLinkDto);
+            var responseContent = CreateValidResponseContent(chargeDto);
             var mockHttpMessageHandler = GetMockHttpMessageHandler(HttpStatusCode.OK, responseContent);
             var httpClient = CreateHttpClient(mockHttpMessageHandler);
             chargesClientFactory.Setup(x => x.CreateClient(httpClient))
@@ -115,9 +115,9 @@ namespace Energinet.DataHub.Charges.Clients.CreateDefaultChargeLink.Tests.Charge
 
             // Assert
             result.Should().NotBeNull();
-            result[0].ChargeId.Should().Be(chargeLinkDto.ChargeId);
-            result[0].ChargeType.Should().Be(chargeLinkDto.ChargeType);
-            result[0].ChargeName.Should().Be(chargeLinkDto.ChargeName);
+            result[0].ChargeId.Should().Be(chargeDto.ChargeId);
+            result[0].ChargeType.Should().Be(chargeDto.ChargeType);
+            result[0].ChargeName.Should().Be(chargeDto.ChargeName);
 
             mockHttpMessageHandler.Protected().Verify(
                 "SendAsync",
@@ -128,7 +128,7 @@ namespace Energinet.DataHub.Charges.Clients.CreateDefaultChargeLink.Tests.Charge
 
         [Theory]
         [InlineAutoDomainData]
-        public async Task GetChargesAsync_WhenResponseIsNotFound_ReturnsEmptyList(
+        public async Task GetChargesAsync_WhenResponseIsNotFound_ReturnsNull(
             Mock<IChargesClientFactory> chargesClientFactory)
         {
             // Arrange
@@ -143,16 +143,123 @@ namespace Energinet.DataHub.Charges.Clients.CreateDefaultChargeLink.Tests.Charge
             var result = await sut.GetChargesAsync().ConfigureAwait(false);
 
             // Assert
-            result.Should().BeOfType<List<ChargeV1Dto>>();
-            result.Should().BeEmpty();
+            result.Should().BeNull();
         }
 
-        private static string CreateValidResponseContent<TModel>(TModel chargeLinkDto)
+        [Theory]
+        [InlineAutoDomainData]
+        public async Task SearchChargesAsync_WhenResponseIsNotFound_ReturnsEmptyList(
+            Mock<IChargesClientFactory> chargesClientFactory,
+            SearchCriteriaV1Dto searchCriteria)
         {
-            var chargeLinks = new List<TModel> { chargeLinkDto };
+            // Arrange
+            var mockHttpMessageHandler = GetMockHttpMessageHandler(HttpStatusCode.NotFound, string.Empty);
+            var httpClient = CreateHttpClient(mockHttpMessageHandler);
+            chargesClientFactory.Setup(x => x.CreateClient(httpClient))
+                .Returns(new ChargesClient(httpClient));
+
+            var sut = chargesClientFactory.Object.CreateClient(httpClient);
+
+            // Act
+            var result = await sut.SearchChargesAsync(searchCriteria).ConfigureAwait(false);
+
+            // Assert
+            result.Should().BeNull();
+        }
+
+        [Theory]
+        [InlineAutoDomainData]
+        public async Task SearchChargesAsync_WhenSuccess_ReturnsCharges(
+            Mock<IChargesClientFactory> chargesClientFactory,
+            SearchCriteriaV1Dto searchCriteria,
+            ChargeV1Dto chargeDto)
+        {
+            // Arrange
+            var responseContent = CreateValidResponseContent(chargeDto);
+            var mockHttpMessageHandler = GetMockHttpMessageHandler(HttpStatusCode.OK, responseContent);
+            var httpClient = CreateHttpClient(mockHttpMessageHandler);
+            chargesClientFactory.Setup(x => x.CreateClient(httpClient))
+                .Returns(new ChargesClient(httpClient));
+
+            var sut = chargesClientFactory.Object.CreateClient(httpClient);
+
+            var expectedUri = new Uri($"{BaseUrl}{ChargesRelativeUris.SearchCharges()}");
+
+            // Act
+            var result = await sut.SearchChargesAsync(searchCriteria).ConfigureAwait(false);
+
+            // Assert
+            result.Should().NotBeNull();
+            result[0].ChargeId.Should().Be(chargeDto.ChargeId);
+            result[0].ChargeType.Should().Be(chargeDto.ChargeType);
+            result[0].ChargeName.Should().Be(chargeDto.ChargeName);
+
+            mockHttpMessageHandler.Protected().Verify(
+                "SendAsync",
+                Times.Exactly(1),
+                ItExpr.Is<HttpRequestMessage>(req => req.Method == HttpMethod.Post && req.RequestUri == expectedUri),
+                ItExpr.IsAny<CancellationToken>());
+        }
+
+        [Theory]
+        [InlineAutoDomainData]
+        public async Task GetMarketParticipantsAsync_WhenSuccess_ReturnsMarketParticipants(
+            MarketParticipantV1Dto marketParticipantDto,
+            Mock<IChargesClientFactory> chargesClientFactory)
+        {
+            // Arrange
+            var responseContent = CreateValidResponseContent(marketParticipantDto);
+            var mockHttpMessageHandler = GetMockHttpMessageHandler(HttpStatusCode.OK, responseContent);
+            var httpClient = CreateHttpClient(mockHttpMessageHandler);
+            chargesClientFactory.Setup(x => x.CreateClient(httpClient))
+                .Returns(new ChargesClient(httpClient));
+
+            var sut = chargesClientFactory.Object.CreateClient(httpClient);
+
+            var expectedUri = new Uri($"{BaseUrl}{ChargesRelativeUris.GetMarketParticipants()}");
+
+            // Act
+            var result = await sut.GetMarketParticipantsAsync().ConfigureAwait(false);
+
+            // Assert
+            result.Should().NotBeNull();
+            result[0].Id.Should().Be(marketParticipantDto.Id);
+            result[0].MarketParticipantId.Should().Be(marketParticipantDto.MarketParticipantId);
+            result[0].Name.Should().Be(marketParticipantDto.Name);
+
+            mockHttpMessageHandler.Protected().Verify(
+                "SendAsync",
+                Times.Exactly(1),
+                ItExpr.Is<HttpRequestMessage>(req => req.Method == HttpMethod.Get && req.RequestUri == expectedUri),
+                ItExpr.IsAny<CancellationToken>());
+        }
+
+        [Theory]
+        [InlineAutoDomainData]
+        public async Task GetMarketParticipantsAsync_WhenResponseIsNotFound_ReturnsNull(
+            Mock<IChargesClientFactory> chargesClientFactory)
+        {
+            // Arrange
+            var mockHttpMessageHandler = GetMockHttpMessageHandler(HttpStatusCode.NotFound, string.Empty);
+            var httpClient = CreateHttpClient(mockHttpMessageHandler);
+            chargesClientFactory.Setup(x => x.CreateClient(httpClient))
+                .Returns(new ChargesClient(httpClient));
+
+            var sut = chargesClientFactory.Object.CreateClient(httpClient);
+
+            // Act
+            var result = await sut.GetMarketParticipantsAsync().ConfigureAwait(false);
+
+            // Assert
+            result.Should().BeNull();
+        }
+
+        private static string CreateValidResponseContent<TModel>(TModel responseDto)
+        {
+            var responseDtos = new List<TModel> { responseDto };
             var options = new JsonSerializerOptions(JsonSerializerDefaults.Web) { Converters = { new JsonStringEnumConverter() } };
 
-            var responseContent = JsonSerializer.Serialize<IList<TModel>>(chargeLinks, options);
+            var responseContent = JsonSerializer.Serialize<IList<TModel>>(responseDtos, options);
             return responseContent;
         }