+++ source/Energinet.Charges.Libraries/source/Energinet.DataHub.Charges.Clients/Charges/ChargesRelativeUris.cs
@@ -37,7 +37,25 @@ namespace Energinet.DataHub.Charges.Clients.Charges
         /// <returns>Relative URI</returns>
         public static Uri GetCharges()
         {
-            return new Uri($"v1/Charges/GetAsync", UriKind.Relative);
+            return new Uri("v1/Charges/GetAsync", UriKind.Relative);
+        }
+
+        /// <summary>
+        /// Provides the relative uri for searching in charges.
+        /// </summary>
+        /// <returns>Relative URI</returns>
+        public static Uri SearchCharges()
+        {
+            return new Uri("v1/Charges/SearchAsync", UriKind.Relative);
+        }
+
+        /// <summary>
+        /// Provides the relative uri for getting all market participants
+        /// </summary>
+        /// <returns>Relative URI</returns>
+        public static Uri GetMarketParticipants()
+        {
+            return new Uri("v1/MarketParticipants/GetAsync", UriKind.Relative);
         }
     }
 }